+++ src/vs/workbench/services/progress/browser/media/progressService.css
