+++ src/vs/workbench/services/progress/test/localProgressService.test.ts
@@ -6,7 +6,7 @@
 import * as assert from 'assert';
 import { IAction, IActionViewItem } from 'vs/base/common/actions';
 import { IEditorControl } from 'vs/workbench/common/editor';
-import { ScopedProgressService, ScopedService } from 'vs/workbench/services/progress/browser/progressService';
+import { ScopedProgressService, ScopedService } from 'vs/workbench/services/progress/browser/localProgressService';
 import { IViewletService } from 'vs/workbench/services/viewlet/browser/viewlet';
 import { IPanelService } from 'vs/workbench/services/panel/common/panelService';
 import { IViewlet } from 'vs/workbench/common/viewlet';
